% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDW.R
\name{IDW}
\alias{IDW}
\title{Inverse distance weighted interpolation (IDW)}
\usage{
IDW(BD_Obs, BD_Coord, shapefile, resolution, idp = 2)
}
\description{
MS-GOP is a machine learning algorithm for merging satellite-based and ground precipitation data.
It combines Random Forest for spatial prediction, residual modeling for bias correction, and quantile mapping for final adjustment, ensuring accurate precipitation estimates across different temporal scales
}
