% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFmerge.R
\name{RFmerge}
\alias{RFmerge}
\title{Merging of satellite datasets with ground observations using RFmerge}
\usage{
RFmerge(
  BD_Obs,
  BD_Coord,
  cov,
  mask = NULL,
  training,
  seed = 123,
  ntree = 2000,
  n_round = NULL,
  Rain_threshold = list(no_rain = c(0, 1)),
  save_model = FALSE,
  name_save = NULL
)
}
\arguments{
\item{BD_Obs}{A \code{data.table} or \code{data.frame} containing observational data with the following structure:
\itemize{
\item The first column (\code{Date}): A \code{Date} object representing the observation date.
\item The remaining columns: Each column corresponds to a unique ground station, where the column name is the station identifier.
}

The dataset should be structured as follows:

\if{html}{\out{<div class="sourceCode">}}\preformatted{> BD_Obs
# A data.table or data.frame with n rows (dates) and m+1 columns (stations + Date)
   Date        ST001  ST002  ST003  ST004  ...
   <date>      <dbl>  <dbl>  <dbl>  <dbl>  ...
1  2015-01-01    0      0      0      0    ...
2  2015-01-02    0      0      0     0.2   ...
3  2015-01-03   0.1     0      0     0.1   ...
}\if{html}{\out{</div>}}
\itemize{
\item Each station column contains numeric values representing observed measurements.
\item The column names (station identifiers) must be unique and match those in \code{BD_Coord$Cod} to ensure proper spatial referencing.
}}

\item{BD_Coord}{A \code{data.table} or \code{data.frame} containing the metadata of the ground stations. It must include the following columns:
#' - \code{"Cod"}:
Unique identifier for each ground station.
\itemize{
\item \code{"X"}:
Latitude of the station in UTM format.
\item \code{"Y"}:
Longitude of the station in UTM format.
\item \code{"Z"}:
Altitude of the station in meters.
}}

\item{cov}{A list of covariates used as independent variables in the RFmerge. Each covariate should be a
\code{SpatRaster} object (from the \code{terra} package) and can represent satellite-derived weather variables or a Digital
Elevation Model (DEM). All covariates should have the same number of layers (bands), except for the DEM, which must have only one layer.}

\item{mask}{A shapefile defining the study area.
If provided, the shapefile must be of class \code{SpatVector} (from the \code{terra} package) with a UTM coordinate reference system.
When specified, a spatial mask is applied to ensure that the final precipitation estimates are restricted to the defined study area.
Defaults to \code{NULL}, meaning no spatial mask is applied.}

\item{training}{Numerical value between 0 and 1 indicating the proportion of data used for model training. The remaining data are used for validation. Note that if you enter, for example, 0.8 it means that 80 \% of the data will be used for training and 20 \% for validation.
If you do not want to perform validation, set training = 1. (Default training = 1).}

\item{seed}{Integer for setting the random seed to ensure reproducibility of results (default: 123).}

\item{ntree}{Numeric indicating the maximum number trees to grow in the Random Forest algorithm. The default value is set to 2000.
This should not be set to too small a number, to ensure that every input row gets predicted at least a few times. If this value is too low, the prediction may be biased.}

\item{n_round}{An integer specifying the number of decimal places to round the interpolated results.
If set to \code{NULL}, all decimal places will be preserved. The default value is \code{1}.}

\item{Rain_threshold}{A list of numeric vectors that define the precipitation thresholds for classifying rainfall events into different categories based on intensity.
Each element of the list should represent a category, with the category name as the list element's name and a numeric vector specifying the lower and upper bounds for that category.}

\item{save_model}{Logical value indicating whether the corrected raster layers should be saved to disk. The default is \code{FALSE}.
If set to \code{TRUE}, make sure to set the working directory beforehand using \code{setwd(path)} to specify where the files should be saved.}

\item{name_save}{Character string. Base name for output file (default: NULL). The output file will be saved as "Model_RFplus.nc".
If you set a different name, make sure you do not set the ".nc" format,
as the code will internally assign it.}
}
\value{
If a value other than 1 is set (point to pixel validation is performed), a list containing two elemeentis returned:

\strong{Ensamble:}
A \code{SpatRaster} object containing the bias-corrected layers for each time step. The number of layers
corresponds to the number of dates for which the correction is applied. This represents the corrected satellite data adjusted for bias.

\strong{Validation:}
A list containing the statistical results obtained from the validation process. This list includes:
\itemize{
\item \code{gof}:
A data table with goodness-of-fit metrics such as Kling-Gupta Efficiency (KGE), Nash-Sutcliffe Efficiency (NSE), Percent Bias (PBIAS), Root Mean Square Error (RMSE), and Pearson Correlation Coefficient (CC). These metrics assess the overall performance of the bias correction process.
\item \code{categorical_metrics}:
A data frame containing categorical evaluation metrics such as Probability of Detection (POD), Success Ratio (SR), False Alarm Rate (FAR), Critical Success Index (CSI), and Hit Bias (HB). These metrics evaluate the classification performance of rainfall event predictions based on user-defined precipitation thresholds.
}

If training is set to 1 (No validation is performed) only the Assembly mentioned above is returned.
}
\description{
RFmerge is a methodology developed by Baez-Villanueva et al. (2020) for the fusion of satellite precipitation datasets with ground-based observations, with the objective of improving the accuracy and spatial representativeness of the data.
This package implements RFmerge using Random Forest as a machine learning technique to correct biases and adjust the distribution of satellite products to in situ measurements.
In addition, it allows the integration of multiple sources of information, including geographic and environmental variables, optimizing the interpolation and spatial extrapolation of precipitation in data-limited regions.
Unlike previous implementations, this package has been optimized to improve computational efficiency and reduce processing times by incorporating advanced data manipulation techniques with \code{data.table}.
}
