% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDW.R
\name{IDW}
\alias{IDW}
\title{Inverse distance weighted interpolation (IDW)}
\usage{
IDW(BD_Obs, BD_Coord, shapefile, resolution, p = 2, n_round = 1)
}
\arguments{
\item{BD_Obs}{A \code{data.table} or \code{data.frame} containing observational data with the following structure:
\itemize{
\item The first column (\code{Date}): A \code{Date} object representing the observation date.
\item The remaining columns: Each column corresponds to a unique ground station, where the column name is the station identifier.
}

The dataset should be structured as follows:

\if{html}{\out{<div class="sourceCode">}}\preformatted{> BD_Obs
# A data.table or data.frame with n rows (dates) and m+1 columns (stations + Date)
   Date        ST001  ST002  ST003  ST004  ...
   <date>      <dbl>  <dbl>  <dbl>  <dbl>  ...
1  2015-01-01    0      0      0      0    ...
2  2015-01-02    0      0      0     0.2   ...
3  2015-01-03   0.1     0      0     0.1   ...
}\if{html}{\out{</div>}}
\itemize{
\item Each station column contains numeric values representing observed measurements.
\item The column names (station identifiers) must be unique and match those in \code{BD_Coord$Cod} to ensure proper spatial referencing.
}}

\item{BD_Coord}{A \code{data.table} or \code{data.frame} containing the metadata of the ground stations. It must include the following columns:
#' - \code{"Cod"}:
Unique identifier for each ground station.
\itemize{
\item \code{"X"}:
Latitude of the station in UTM format.
\item \code{"Y"}:
Longitude of the station in UTM format.
}}

\item{shapefile}{A shapefile defining the study area, used to constrain the interpolation to the region of interest.
The shapefile must be of class \code{SpatVector} (from the \code{terra} package) and should have a UTM coordinate reference system.}

\item{resolution}{A numeric value indicating the resolution of the interpolation grid in kilometers (\code{km}).}

\item{p}{'Numeric' value that controls how the influence decreases with distance. The default value is 2}

\item{n_round}{An integer specifying the number of decimal places to round the interpolated results.
If set to \code{NULL}, all decimal places will be preserved. The default value is \code{1}.}
}
\value{
A \code{SpatRaster} object (from the \code{terra} package) where:
\itemize{
\item Each layer (\verb{raster band}) corresponds to the interpolated values for a specific date in \code{BD_Obs$Date}.
\item The spatial resolution is defined by the \code{resolution} parameter (in km).
\item The coordinate reference system (CRS) matches that of the input \code{shapefile}.
}
}
\description{
This function performs Inverse Distance Weighting (IDW), which is a spatial interpolation method that estimates unknown values using the known values of surrounding points, assigning greater influence to closer points.
}
\section{Details}{

Inverse distance weighting (IDW) works as a deterministic mathematical interpolator that assumes
that values closer to an unknown point are more closely related to it than those farther away. When using this method, sample points are weighted during the interpolation process so that the influence of a known point on the unknown point decreases as the distance between them increases.
The IDW method calculates the value at an unknown point by a weighted average of the values of the known points, where the weights are inversely proportional to the distances between the prediction point and the known points. The basic formula defined by Shepard states that:
\deqn{\hat{Z}(s_0) = \frac{\sum_{i=1}^{N} w_i Z(s_i)}{\sum_{i=1}^{N} w_i}}
where:
\describe{
\item{\eqn{\hat{Z}(s_0)}}{ is the estimated value at the unknown point.}
\item{\eqn{Z(s_i)}}{ are the known values.}
\item{\eqn{w_i}}{ are the weights assigned to each known point.}
\item{\eqn{N}}{ is the total number of known points used.}
}

The weights are calculated by:
\deqn{w_i = \frac{1}{d(s_0, s_i)^p}}
where:
\describe{
\item{\eqn{d(s_0, s_i)}}{ is the distance between the unknown point \eqn{s_0} and the known point \eqn{s_i}.}
\item{\eqn{p}}{ is the power parameter that controls how the influence decreases with distance.}
}
}

\examples{
\donttest{
library(InterpolateR)
# Load data from on-site observations
 data("BD_Obs", package = "InterpolateR")
 data("BD_Coord", package = "InterpolateR")

# Load the study area where the interpolation is performed.
 shapefile <- terra::vect(system.file("extdata/study_area.shp", package = "InterpolateR"))

 # Perform the interpolation
 Interpolated_data <- IDW(BD_Obs, BD_Coord, shapefile, resolution = 5, p = 2, n_round = 1)
}
}
\references{
Shepard, D. (1968) A two-dimensional interpolation function for irregularly-spaced data. Proceedings of the 1968 ACM National Conference, 1968, pages 517--524. DOI: 10.1145/800186.810616
}
\author{
Jonnathan Landi \href{mailto:jonnathan.landi@outlook.com}{jonnathan.landi@outlook.com}
}
